% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainogram.R
\name{calc.Qi}
\alias{calc.Qi}
\title{Rank based scores}
\usage{
calc.Qi(x)
}
\arguments{
\item{x}{vector of numeric values.}
}
\value{
calc.Qi(x) = (rank(x) - 0.5) / n\cr where n is the length of x and ranking
is performed in decreasing order:
\itemize{
   \item rank(min(x)) = n
   \item rank(max(x)) = 1
}
}
\description{
Transform numeric values into rank-based scores.
}
\seealso{
\link{enrichmentScore},
   \link{domainogram},
   \link{segmentation}
}
\author{
Benjamin Leblanc
}
