% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.functions.R
\name{colorstrip}
\alias{colorstrip}
\title{colorstrip}
\usage{
colorstrip(Pw, x, y, width, height, colfun = makeColors, n.ticks = 4,
  n.bins = 100, cex = 0.5, ...)
}
\arguments{
\item{Pw}{}

\item{x}{}

\item{y}{}

\item{width}{}

\item{height}{}

\item{colfun}{}

\item{n.ticks}{}

\item{n.bins}{}

\item{cex}{}

\item{\dots}{}
}
\author{
Benjamin Leblanc
}
