% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restriction.map.R
\name{combineByFragment}
\alias{combineByFragment}
\title{Combine probe measurements by restriction fragments}
\usage{
combineByFragment(x, probes.grg, FUN = mean, ...)
}
\arguments{
\item{x}{}

\item{probes.grg}{}

\item{FUN}{}

\item{\dots}{}
}
\description{

}
\seealso{
\link{}
}
\author{
Benjamin Leblanc
}
