% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restriction.map.R
\name{countProbesPerFragment}
\alias{countProbesPerFragment}
\title{Count the number of accepted probes per restriction fragment}
\usage{
countProbesPerFragment(probes.grg)
}
\arguments{
\item{probes.grg}{}
}
\description{

}
\seealso{
\link{matchProbesToFragments},
   \link{matchFragmentsToProbes},
   \link{computeRestrictionMap}
}
\author{
Benjamin Leblanc
}
