% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainogram.R
\name{domainogram}
\alias{domainogram}
\title{Domainogram visualization}
\usage{
domainogram(
  Yi,
  x.s,
  x.e,
  w2y,
  xlim = NULL,
  wlim = NULL,
  colfun = makeColors,
  xlab = "",
  add = F,
  logskip = 1.01,
  ...
)
}
\arguments{
\item{Yi}{vector of statistical scores produced by the \link{enrichmentScore} function}

\item{x.s}{vector of start coordinates (x axis) defining the represented genomic
intervals.}

\item{x.e}{vector of end coordinates (x axis) defining the represented genomic
intervals.}

\item{w2y}{coordinate mapping for window sizes (y axis) produced by the
\link{wSize2yAxis} function.}

\item{xlim}{range of the represented genomic region, which must be indicated as
c(xmin, xmax).}

\item{wlim}{range of represented window sizes, which must be indicated as c(wmin, wmax).}

\item{colfun}{coloring function. The function \link{makeColors} is used by default.}

\item{xlab}{legend of the x axis.}

\item{add}{logical, when set to TRUE the plot should overlay an existing multi-resolution
plot.}

\item{logskip}{numeric coefficient for skipping the representation of window sizes when
using the log2 scale on the y axis.
A value of 1.0 forces to represent all window sizes.
A value of 2 would double the window sizes represented at each iteration of
the rendering.}

\item{\dots}{optional parameters forwarded to the \code{colfun} function.
When using the default \link{makeColors} function, these parameters are
\code{thresholds}, \code{colors}, \code{background} and \code{overflow}
(see documentation for details).}
}
\description{
Plot a domainogram from a sequence of log-transformed rank-based scores.
This function allows to use a linear or log2 scale for window sizes (y
axis).
}
\details{
The domainogram is a multi-resolution representation of scan statistics
introduced by de Wit et al., 2008 (see references below). It is based on
R.A. Fisher's method for the combination of p-values applied to
non-parametric scores.
}
\examples{
# Simulate enrichment signal
n <- 500
Mi <- cos(2*pi*2.5*(-n:n)/n) + rnorm(2*n+1)
n <- length(Mi)

# Compute enrichment scores
Yi <- enrichmentScore(Mi)

# Visualization coordinates
x.s <- 1:n - 0.5
x.e <- 1:n + 0.5
w2y <- wSize2yAxis(n)

layout(matrix(1:2, 2, 1), heights=c(3,1)/4)
par(mar=c(3, 4, 1, 2)) # default bottom, left, top, right = c(5, 4, 4, 2)

# Plot domainogram
domainogram(Yi, x.s, x.e, w2y)
plot(Mi, type='l')
}
\references{
de Wit E., Braunschweig U., Greil F., Bussemaker H.J., and van Steensel B.
Global chromatin domain organization of the Drosophila genome.
PLoS genetics 4: e1000045 (2008).
\url{http://dx.doi.org/10.1371/journal.pgen.1000045}
}
\seealso{
\link{calc.Qi},
   \link{enrichmentScore},
   \link{visualizationCoordinates},
   \link{wSize2yAxis},
   \link{makeColors},
   \link{segmentation},
   \link{plotOptimalSegments},
   \link{plotDomains}
}
\author{
Benjamin Leblanc
}
