% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{enrichmentQuality}
\alias{enrichmentQuality}
\title{Best and worst enrichment levels}
\usage{
enrichmentQuality(x, y, q.best = 0.015, q.worst = 0.005, plots = F)
}
\arguments{
\item{x}{raw Ctr signal intensity}

\item{y}{raw 4C signal intensity}

\item{q.best}{}

\item{q.worst}{}

\item{plots}{}
}
\description{
Identify probes most likely associated with best and worst enrichment levels.
}
\examples{
x <- 2^pmin(rlnorm(100000, 1, 0.5), 10)
y <- 2^pmin(rlnorm(100000, 0.5, 1), 10)

chk <- enrichmentQuality(x, y, plots=T)

# Visualize thresholds on enrichment quality scores
clr <- with(
  chk, rgb(is.worst, is.best, 0, ifelse(is.worst | is.best, 0.5, 0.1))
)
with(chk, plot(w.score, b.score, pch='.', col=clr))
}
\author{
Benjamin Leblanc
}
\seealso{
\link{plotProbeDistanceControls}
}

