% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainogram.R
\name{enrichmentScore}
\alias{enrichmentScore}
\title{Log-transformed enrichment scores}
\usage{
enrichmentScore(x)
}
\arguments{
\item{x}{vector of numeric values}
}
\value{
log(calc.Qi(x))
}
\description{
Compute log-transformed rank-based scores from numeric values.
}
\seealso{
\link{calc.Qi},
   \link{domainogram},
   \link{segmentation}
}
\author{
Benjamin Leblanc
}
