% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtree.R
\name{kdtree.getDN}
\alias{kdtree.getDN}
\title{Neighbor selection}
\usage{
kdtree.getDN(M, v, d, node)
}
\arguments{
\item{M}{}

\item{v}{}

\item{d}{}

\item{node}{}
}
\description{
Neighbor selection based on a pre-computed kd-tree structure of (A,M)
values.
}
\author{
Benjamin Leblanc
}
