% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtree.R
\name{kdtree}
\alias{kdtree}
\title{kd-tree representation of (A,M) values}
\usage{
kdtree(M, i, d)
}
\arguments{
\item{M}{matrix of A and M values}

\item{i}{row indexes}

\item{d}{}
}
\description{
generate a kd-tree representation of (A,M) values for faster execution of the
background traversal algorithm (see normalizeArrayData).
}
\author{
Benjamin Leblanc
}
