% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localopt.c.R
\name{localopt.c}
\alias{localopt.c}
\title{local optimization}
\usage{
localopt.c(Yi, name, wmax = 0, wmin = 0, gamma = 1, Tw = 0)
}
\description{

}
\seealso{
\link{domainogram},
}
\author{
Benjamin Leblanc
}
