% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{lowessCorrection}
\alias{lowessCorrection}
\title{LOWESS based normalization}
\usage{
lowessCorrection(x, y, lowess.f = 0.3, lowess.mad = 0, lowess.iter = 5,
  plots = F)
}
\arguments{
\item{x}{}

\item{y}{}

\item{lowess.f}{}

\item{lowess.mad}{}

\item{lowess.iter}{}

\item{plots}{}
}
\description{

}
\author{
Benjamin Leblanc
}
\seealso{
\link{normalizeArrayData}
}

