% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeColors.R
\name{makeColors}
\alias{makeColors}
\title{Represent numerical values by a color vector}
\usage{
makeColors(p, thresholds = c(10, 50, 500, 5000), colors = c("blue",
  "yellow", "red", "black"), background = "white", overflow = "white")
}
\arguments{
\item{p}{vector of positive numeric values.}

\item{thresholds}{numeric vector defining the boundaries between ranges of values to be
mapped into a specific color interval (default = c(10, 50, 500, 5000)).}

\item{colors}{color vector providing the colors representing each threshold value
(default = c("blue", "yellow", "red", "black")).}

\item{background}{color representing p = 0, which is set to "white" by default.}

\item{overflow}{color representing p > max(thresholds), which is set to "white" by default.}
}
\value{
makeColors returns a character vector of RGBA colors in hexadecimal.
}
\description{
makeColors builds a vector of RGB colors to represent numerical values.
This color representation is generated by applying piecewise color mapping
to input values.
}
\author{
Benjamin Leblanc
}
\seealso{
\link{domainogram}
}

