% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanshift.R
\name{mean.shift}
\alias{mean.shift}
\title{Meanshift}
\usage{
\method{mean}{shift}(
  M,
  v,
  d,
  node,
  epsilon = 0.001,
  col = rgb(1, 0.5, 0, 1),
  nsteps = 0,
  plots = FALSE
)
}
\description{
Core function of the background traversal algorithm based on a pre-computed
kd-tree structure of (A,M) values. See normalizeArrayData function.
}
\seealso{
\link{},
}
\author{
Benjamin Leblanc
}
