\name{MRA.TA-package}
\alias{MRA.TA-package}
\alias{MRA.TA}
\docType{package}
\title{
\packageTitle{MRA.TA}
}
\description{
\packageDescription{MRA.TA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MRA.TA}
\packageIndices{MRA.TA}
%% ~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MRA.TA}

Maintainer: \packageMaintainer{MRA.TA}
}
\references{
%% ~~ Literature or other references for background information ~~
}
%% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
