% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainogram.R
\name{new.domain.plot}
\alias{new.domain.plot}
\title{Create a new multi-resolution plot}
\usage{
new.domain.plot(x.s, x.e, w2y, xlim = NULL, wlim = NULL, xlab = "")
}
\arguments{
\item{x.s}{}

\item{x.e}{}

\item{w2y}{}

\item{xlim}{}

\item{wlim}{}

\item{xlab}{}
}
\description{

}
\seealso{
\link{domainogram},
   \link{plotOptimalSegments}
}
\author{
Benjamin Leblanc
}
