% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeArrayData.R
\name{normalizeArrayData}
\alias{normalizeArrayData}
\title{Normalize tiling array data}
\usage{
normalizeArrayData(A, M, smoothness = 0.08, epsilon = 0.01, nsteps = 11,
  name = "Test", plots = TRUE, lowess = T, lowess.f = 0.2,
  lowess.mad = 0, lowess.iter = 5)
}
\arguments{
\item{A}{}

\item{M}{}

\item{smoothness}{}

\item{epsilon}{}

\item{nsteps}{}

\item{name}{}

\item{plots}{}

\item{lowess}{}

\item{lowess.f}{}

\item{lowess.mad}{}

\item{lowess.iter}{}
}
\description{

}
\examples{
# Load array data and apply background bias estimation+correction, followed by
# lowess normalization

data(WT.4C.Fab7.dm6)

# 1. Combined procedure ------------------------------------------------------

# Raw A and M values
A <- (log2(r1.4C$PM) + log2(r1.ct$PM))/2
M <- (log2(r1.4C$PM) - log2(r1.ct$PM))

# Normalized A and M values
res <- normalizeArrayData(
  A, M, name="4C_norm", plots=TRUE
)
A <- res$A; M <- res$M

# 2. Equivalent step by step procedure ---------------------------------------

# Raw A and M values
A <- (log2(r1.4C$PM) + log2(r1.ct$PM))/2
M <- (log2(r1.4C$PM) - log2(r1.ct$PM))

# Estimate background bias
bb.r1 <- backgroundBiasEstimation(A, M, plots = T)

# Correct background bias
res <- backgroundBiasCorrection(A, M, theta=bb.r1)
A <- res$x; M <- res$y

# Apply lowess normalization
res <- lowessCorrection(A, M, lowess.f=0.2, plots = T)
A <- res$x; M <- res$y
}
\seealso{
\link{backgroundBiasEstimation},
   \link{backgroundBiasCorrection},
   \link{lowessCorrection}
}
\author{
Benjamin Leblanc
}
