% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProbeDistanceControls.R
\name{plotProbeDistanceControls}
\alias{plotProbeDistanceControls}
\title{Plot 4C enrichment versus distance to restriction sites}
\usage{
plotProbeDistanceControls(
  x,
  rnk,
  dis,
  dlim = c(-2000, 2000),
  QF = NULL,
  n.q = 100,
  ylab = "4C"
)
}
\arguments{
\item{x}{}

\item{rnk}{}

\item{dis}{}

\item{dlim}{}

\item{QF}{}

\item{n.q}{}

\item{ylab}{}
}
\description{
Plot 4C enrichment versus distance to restriction sites
}
\seealso{
\link{plotSelectedProbes},
   \link{enrichmentQuality}
}
\author{
Benjamin Leblanc
}
