% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSelectedProbes.R
\name{plotSelectedProbes}
\alias{plotSelectedProbes}
\title{Plot 4C enrichment versus distance to restriction sites}
\usage{
plotSelectedProbes(a, m, dis, sel, dlim = c(-2000, 2000), ylab = "4C")
}
\description{

}
\seealso{
\link{plotProbeDistanceControls},
   \link{enrichmentQuality}
}
\author{
Benjamin Leblanc
}
