% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processArgs.R
\name{processArgs}
\alias{processArgs}
\title{Extract command line arguments}
\usage{
processArgs(script.args = NULL, auto.help = T, extra.help = "",
  verbose = T)
}
\arguments{
\item{script.args}{matrix or data.frame with argument specifications}

\item{auto.help}{}

\item{extra.help}{}

\item{verbose}{}
}
\description{

}
\author{
Benjamin Leblanc
}
