% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Read tabulated text files}
\usage{
readData(fpath, fform)
}
\arguments{
\item{fpath}{}

\item{fform}{file format definition (see details below).}
}
\value{
readData returns a data.frame
}
\description{
Read tab delimited text files with specific format definition
}
\details{
List of available format definitions for tab delimited text files:
\itemize{
  \item \code{experimental.design}
  \item \code{mapped.probe.sequences}
  \item \code{restriction.fragment}
  \item \code{restriction.site}
  \item \code{nimblegen.ndf}
  \item \code{nimblegen.updated}
  \item \code{nimblegen.pair}
}

Each format definition provides the following informations:
\itemize{
  \item \code{comment} : \code{character}, same as the \code{comment.char} parameter of \code{read.table} function
  \item\code{skip}     : \code{integer}, same as the \code{skip} parameter of \code{read.table} function
  \item\code{header}   : \code{logical}, same as the \code{header} parameter of \code{read.table} function
  \item\code{sep}      : \code{character}, same as the \code{sep} parameter of \code{read.table} function
  \item\code{table}    : \code{data.frame} with the following fields
  \itemize{
    \item\code{columns}  : \code{character}, original column names when header is present
    \item\code{classes}  : \code{character}, R classes used to import column values. Same as the \code{colClasses} parameter of \code{read.table} function
    \item\code{extract}  : \code{logical}, enable/disable the importation of each column
    \item\code{name.as}  : \code{character}, optional renaming of extracted columns
    \item\code{description} : \code{character}, column description
  }
}
}
\seealso{
\link{verifyInputFiles}
}
\author{
Benjamin Leblanc
}
