% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.formats.R
\name{verifyInputFiles}
\alias{verifyInputFiles}
\title{Check input files}
\usage{
verifyInputFiles(path.list, columns = NULL, sep = "\\t", comments = "#")
}
\arguments{
\item{path.list}{}

\item{columns}{}

\item{sep}{}

\item{comments}{}
}
\description{
Check availability of tabulated text input files, including basic
verification of the expected format.
}
\author{
Benjamin Leblanc
}
\seealso{
\link{readData}
}

