% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainogram.R
\name{visualizationCoordinates}
\alias{visualizationCoordinates}
\title{Continuous coordinates for domainogram visualization (x axis)}
\usage{
visualizationCoordinates(start, end)
}
\arguments{
\item{start}{vector defining the start positions of genomic intervals.}

\item{end}{vector defining the end positions of genomic intervals.}
}
\value{
\item{start }{start positions for domainogram visualization}
\item{end }{end positions for domainogram visualization}
}
\description{
Compute coordinates for the domainogram visualization of a range of genomic
positions (x axis).
}
\author{
Benjamin Leblanc
}
\seealso{
\link{domainogram},
   \link{wSize2yAxis}
}

