% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainogram.R
\name{wSize2yAxis}
\alias{wSize2yAxis}
\title{Continuous coordinates for domainogram visualization (y axis)}
\usage{
wSize2yAxis(wmax, logscale = FALSE)
}
\arguments{
\item{wmax}{integer indicating the maximum window size to be represented}

\item{logscale}{logical indicating if window sizes should be represented according to a log2
scale instead of a linear scale (default = F, linear)}
}
\value{
\item{logscale}{value of the \code{logscale} parameter}
\item{ylab}{text to be used for the y axis legend}
\item{y.t}{top coordinate for each window size}
\item{y.b}{bottom coordinate for each window size}
\item{ylim}{range of the visualization coordinates (y axis)}
}
\description{
Compute coordinates for the domainogram visualization of a range of window
sizes (y axis).
}
\seealso{
\link{domainogram},
   \link{visualizationCoordinates},
}
\author{
Benjamin Leblanc
}
