% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{backgroundBiasCorrection}
\alias{backgroundBiasCorrection}
\title{Background bias correction}
\usage{
backgroundBiasCorrection(x, y, theta, AM.scale.compensation = T)
}
\arguments{
\item{x}{numeric vector of A values, e.g. the average of log2 of the specific and the
reference signals.}

\item{y}{numeric vector of M values, e.g. the log2 ratio of the specific over the
reference signal.}

\item{theta}{estimated bias angle in radians.}

\item{AM.scale.compensation}{\code{logical} determining if A and M values should be rescaled prior to
correction by a rotation of angle \code{-theta}.
Scale compensation is required and activated by default.}
}
\value{
\code{backgroundBiasCorrection} returns a \code{list} with the following
elements:
\item{x}{ corrected A value}
\item{y}{ corrected M value}
}
\description{
Correct for a background bias in A, M values from a tiling array.
The correction consists in a rotation of angle \code{-theta}, as estimated by the
\link{backgroundBiasEstimation} function.
}
\author{
Benjamin Leblanc
}
\seealso{
\link{backgroundBiasEstimation},
   \link{normalizeArrayData}
}

