% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRestrictionMap.R
\name{computeRestrictionMap}
\alias{computeRestrictionMap}
\title{Compute a restriction map}
\usage{
computeRestrictionMap(genome.seq, enzyme.motif, output.file, seqlist = NULL)
}
\arguments{
\item{genome.seq}{\link{DNAStringSet} object containing the genome sequence}

\item{enzyme.motif}{\code{character} defining the DNA sequence of enzyme cut sites}

\item{output.file}{\code{character} prefix used to generate output file names.}

\item{seqlist}{indices of chromosomes to be considered (default = all).}
}
\value{
\item{sites}{file name with the restriction sites data}
\item{fragments}{file name with restriction fragments data}
}
\description{

}
\seealso{
\link{matchFragmentsToProbes},
   \link{matchProbesToFragments},
   \link{DNAStringSet}
}
\author{
Benjamin Leblanc
}
