% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restriction.map.R
\name{matchFragmentsToProbes}
\alias{matchFragmentsToProbes}
\title{Match restriction fragments to microarray probes}
\usage{
matchFragmentsToProbes(fragments.grg, probes.grg)
}
\arguments{
\item{fragments.grg}{}

\item{probes.grg}{}
}
\description{

}
\author{
Benjamin Leblanc
}
\seealso{
\link{matchProbesToFragments},
   \link{computeRestrictionMap}
}

