% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchProbesToFragments.R
\name{matchProbesToFragments}
\alias{matchProbesToFragments}
\title{Match micro array probes to restriction fragments}
\usage{
matchProbesToFragments(probes.grg, fragments.grg)
}
\arguments{
\item{probes.grg}{}

\item{fragments.grg}{}
}
\description{

}
\seealso{
\link{matchFragmentsToProbes},
   \link{computeRestrictionMap}
}
\author{
Benjamin Leblanc
}
