% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtree.R
\name{plot.kdtree.border}
\alias{plot.kdtree.border}
\title{visualize the kd-tree structure}
\usage{
plot.kdtree.border(node)
}
\description{
Utility function for visualizing the kd-tree structure of (A,M) values.
}
\author{
Benjamin Leblanc
}

